+++ src/MultiTranslator.php
@@ -7,6 +7,7 @@ use craft\base\Element;
 use craft\base\Model;
 use craft\base\Plugin;
 use craft\elements\Asset;
+use craft\elements\Category;
 use craft\elements\Entry;
 use craft\events\DefineHtmlEvent;
 use craft\events\RegisterElementActionsEvent;
@@ -261,6 +262,7 @@ class MultiTranslator extends Plugin
     {
         $supportedElementClasses = [
             Entry::class,
+            Category::class,
             Asset::class,
             'craft\commerce\elements\Product',
             'craft\commerce\elements\Variant',